requirements.txt
transformers==4.38.0
accelerate==0.24.1
pillow